phyloTAGs description
The phyloTAGs pipeline is a set of PERL scripts that assist in the
selection of low-degeneracy primers used to amplify single-copy genes
from bacteria representing a set of taxonomic groups, selected
according to their evolutionary relatedness (i.e, their level of 16s
rDNA divergence). The scripts expedite the process of alignment, search
and assessment of polymorphic regions within the targeted
protein-coding gene. The phyloTAGs pipeline is composed of the
following scripts: (1) 01_phyloTAGs_align.pl and (2)
02_phyloTAGs_primer.pl, which, respectively, align genes by codons and
design primers along the alignment.